Dear God...
git-svn-id: file:///home/nbingham/backup/private-repos/trunk/Software/Work/chp@276 404bc64d-9d1b-498e-bf1b-9a6e5dac48b1

+++ src/data/sspace.cpp
@@ -9,7 +9,7 @@
 
 #include "sspace.h"
 
-int state_space::size()
+size_t state_space::size()
 {
 	return states.size();
 }
